refactor: change dir of UserContext
